Introduce xmlpp::wrapped_exception
This is an internal class which is used by SaxParser and Validator to
save the exception object thrown by a handler method when the exception
does not derive from xmlpp::exception (e.g. std::exception). The Raise()
method of xmlpp::wrapped_exception calls std::rethrow_exception() to
rethrow the exception object thrown by the handler method.

Catching any exception object thrown by a handler method is important in
ensuring that we are able to reset the internal state, and, in the case
of SaxParser::parse(), that we restore the old _xmlSAXHandler pointer so
that we do not double-free the _xmlSAXHandler object held by SaxParser.

Fixes Bug 753570 - “double free or corruption” if a std::exception is thrown
https://bugzilla.gnome.org/show_bug.cgi?id=753570