Add NMake Makefiles
This adds a set of NMake Makefiles that can be used to build libxml++,
which will replace the Visual Studio 2010(!) projects that reside in the
repository.

To be consistent with the rest of the -mm stack that we build with Visual
Studio, we will need to separate the output filenames for the resulting .lib
and .dll files for Visual Studio 2017 and 2019.  A NMake Makefile build option
"USE_COMPAT_LIBS" is added if building 'xml++-vc150-x_y.[dll|lib] is still desired.

Note that this branch does not rely on glibmm nor libsigc++, so linking against
libxml2 built with an older Visual Studio version should work properly, if
libxml2 is built with Visual Studio 2015 or later, but again, we are enforcing
the Visual Studio toolset versioning policy by default for consistency's sake.