meson.build: Also look for libxml2 manually on Windows
The Windows-specific Makefiles for libxml2 do not generate pkg-config files for
us, so we must look for it manually on Windows if the libxml2 pkg-config file
does not exist.

This will check for the library and the headers, and run a simple version check
to make sure things are indeed in place.

Also make the build look for glibmm only via pkg-config, since we have Meson build
files for glibmm already

Also make sure that we have the correct info put into the pkg-config file that we
generate