NMake Makefiles: Fix build from tarballs
Update the rules so that we won't attempt to unnecessarily generate
libxml++config.h and libxml++.rc in our builds, and make sure that we do indeed
generate them with the version info if generating them becomes necessary.

Also streamline this process so that it will be part of the 'all' target.