NMake Makefiles: Make things more configurable
Add the INCLUDEDIR and LIBDIR options, so that we can set where we find
dependent headers and libraries, if they are not in %INCLUDE% and/or %LIB%
and are not in $(PREFIX)\include and $(PREFIX)\lib.

Also mention about the PREFIX and PERL options, so that one can use them when
necessary.