exceptions/exception.h: Try fixing linking on MinGW
Using the xmlpp::exception class may lead to a linker error on its
vtable on MinGW, unless the whole class is exported, but we also want to
avoid exporting that whole class under Visual Studio, so that we can
avoid having the built DLL/import library being tied to a particular STL
version, so define macros to accomdate such situations and apply it to
the xmlpp::exception class.

This should not affect the ABI stability on Visual Studio builds.

Fixes: https://github.com/libxmlplusplus/libxmlplusplus/issues/45