Meson: Cleanup and fix libxml2 dep search
Use CMake to help us find libxml2, in addition to using pkg-config, as
CMake actually looks up libxml2 for us using the headers in %INCLUDE%
and the libraries in %LIB% in a more comprehensive way, so we don't need
to reinvent the wheel.

Since we use Meson 0.60.0 or later, we can use the CMake libxml2
dependency name together with libxml-2.0.pc to find libxml2 in one
single call.

Also update the libxml2 CMake subproject by using CMake to find ICU and
liblzma in the same manner, and ensure that we have the correct
libxml2 .lib in the resulting pkg-config file for libxml++.